fix(templates): fix displayed version in description
Related to camunda/camunda-modeler#2053
