+++ lib/provider/camunda/element-templates/parts/DescriptionProps.js
@@ -1,7 +1,6 @@
 var escapeHTML = require('../../../../Utils').escapeHTML;
 
-var getTemplateId = require('../Helper').getTemplateId,
-    getTemplateVersion = require('../Helper').getTemplateVersion;
+var getTemplateId = require('../Helper').getTemplateId;
 
 var domClasses = require('min-dom').classes,
     domEvent = require('min-dom').event,
@@ -11,6 +10,8 @@ var domClasses = require('min-dom').classes,
 var isUndefined = require('lodash/isUndefined');
 
 var getBusinessObject = require('bpmn-js/lib/util/ModelUtil').getBusinessObject;
+
+var getVersionOrDateFromTemplate = require('./Helper').getVersionOrDateFromTemplate;
 
 var MAX_DESCRIPTION_LENGTH = 200;
 
@@ -53,12 +54,10 @@ module.exports = function(
     });
   }
 
-  var version = getVersionOrDateFromElement(element);
-
-  if (version) {
+  if (currentElementTemplate) {
     entries.push({
       id: 'element-template-version',
-      html: '<p>' + translate('Version') + ': ' + version + '</p>'
+      html: '<p>' + getVersionOrDateFromTemplate(currentElementTemplate) + '</p>'
     });
   }
 
@@ -244,53 +243,4 @@ function findNewestElementTemplate(elementTemplates, currentElementTemplate) {
 
       return newestElementTemplate;
     }, null);
-}
-
-function leftPad(string, length, character) {
-  while (string.length < length) {
-    string = character + string;
-  }
-
-  return string;
-}
-
-function getVersionOrDateFromElement(element) {
-  var version = getTemplateVersion(element);
-
-  if (isUndefined(version)) {
-    return null;
-  }
-
-  return 'Version ' + (toDateString(version) || version);
-}
-
-function getVersionOrDateFromTemplate(template) {
-  var metadata = template.metadata,
-      version = template.version;
-
-  if (metadata) {
-    if (!isUndefined(metadata.created)) {
-      return 'Version ' + toDateString(metadata.created);
-    } else if (!isUndefined(metadata.updated)) {
-      return 'Version ' + toDateString(metadata.updated);
-    }
-  }
-
-  if (isUndefined(version)) {
-    return null;
-  }
-
-  return 'Version ' + version;
-}
-
-function toDateString(timestamp) {
-  var date = new Date(timestamp);
-
-  var year = date.getFullYear();
-
-  var month = leftPad(String(date.getMonth() + 1), 2, '0');
-
-  var day = leftPad(String(date.getDate()), 2, '0');
-
-  return day + '.' + month + '.' + year;
 }
