+++ lib/provider/camunda/element-templates/parts/Helper.js
@@ -1,6 +1,7 @@
 'use strict';
 
-var find = require('lodash/find');
+var find = require('lodash/find'),
+    isUndefined = require('lodash/isUndefined');
 
 var isAny = require('bpmn-js/lib/features/modeling/util/ModelingUtil').isAny;
 
@@ -74,4 +75,43 @@ function option(value, template, translate) {
   }
 
   return option;
+}
+
+module.exports.getVersionOrDateFromTemplate = function(template) {
+  var metadata = template.metadata,
+      version = template.version;
+
+  if (metadata) {
+    if (!isUndefined(metadata.created)) {
+      return 'Version ' + toDateString(metadata.created);
+    } else if (!isUndefined(metadata.updated)) {
+      return 'Version ' + toDateString(metadata.updated);
+    }
+  }
+
+  if (isUndefined(version)) {
+    return null;
+  }
+
+  return 'Version ' + version;
+};
+
+function toDateString(timestamp) {
+  var date = new Date(timestamp);
+
+  var year = date.getFullYear();
+
+  var month = leftPad(String(date.getMonth() + 1), 2, '0');
+
+  var day = leftPad(String(date.getDate()), 2, '0');
+
+  return day + '.' + month + '.' + year;
+}
+
+function leftPad(string, length, character) {
+  while (string.length < length) {
+    string = character + string;
+  }
+
+  return string;
 }
